requirements.txt
flask==2.3.3
flask-dance==7.0.0
requests==2.31.0
python-dotenv==1.0.0

